+++ src/routes/index.tsx
@@ -2,7 +2,10 @@ import { Suspense, lazy } from "react";
 import { Navigate, useRoutes, useLocation } from "react-router-dom";
 import Loading from "../components/Loading";
 import MainLayout from "../pages/MainLayout";
-
+//guards
+import AuthGuard from "../guards/AuthGuard";
+import RoleBasedGuard from "../guards/RoleBasedGuard";
+import { PATH_MODULE } from "./path";
 // ----------------------------------------------------------------------
 
 const Loadable = (Component: React.ElementType) => (props: any) => {
@@ -33,14 +36,34 @@ export default function Router() {
     },
     {
       path: "module",
-      element: <MainMenu />,
+      element: (
+        // <AuthGuard  element={<MainMenu />}/>
+        <AuthGuard>
+          <Module />
+        </AuthGuard>
+      ),
     },
     {
       path: "module",
-      element: <MainLayout />,
+      element: (
+        // <AuthGuard  element={<MainLayout />}/>
+        <AuthGuard>
+          <MainLayout />
+        </AuthGuard>
+      ),
       children: [
         { element: <Navigate to="/module/cashier" replace /> },
-        { path: "cashier", element: <Cashier /> },
+        {
+          path: "cashier",
+          element: (
+            <RoleBasedGuard
+              accessibleRoles={["Admin"]}
+              prevPath={PATH_MODULE.root}
+            >
+              <Cashier />
+            </RoleBasedGuard>
+          ),
+        },
         { path: "accounting", element: <Accounting /> },
         { path: "admission", element: <Admission /> },
         { path: "campuses", element: <Campuses /> },
@@ -58,10 +81,24 @@ export default function Router() {
               path: "student-type",
               children: [
                 { path: "", element: <StudentType /> },
-                { path: "add", element: <CreateStudentType /> },
+                {
+                  path: "add",
+                  element: (
+                    <RoleBasedGuard
+                      accessibleRoles={["Admin"]}
+                      prevPath={PATH_MODULE.settings.studentType.root}
+                    >
+                      <CreateStudentType />{" "}
+                    </RoleBasedGuard>
+                  ),
+                },
               ],
             },
           ],
+        },
+        {
+          path: "semester",
+          children: [{ path: "create", element: <CreateSemester /> }],
         },
       ],
     },
@@ -73,7 +110,7 @@ export default function Router() {
 
 // Authentication
 const Login = Loadable(lazy(() => import("../pages/Login")));
-const MainMenu = Loadable(lazy(() => import("../pages/MainMenu")));
+const Module = Loadable(lazy(() => import("../pages/Module")));
 const Registrar = Loadable(lazy(() => import("../pages/Registrar")));
 const DefaultComponents = Loadable(
   lazy(() => import("../pages/DefaultComponents"))
@@ -92,6 +129,9 @@ const StudentType = Loadable(
 );
 const CreateStudentType = Loadable(
   lazy(() => import("../pages/Settings/CreateStudentType"))
+);
+const CreateSemester = Loadable(
+  lazy(() => import("../pages/Semester/CreateSemester"))
 );
 // Main
 // const ComingSoon = Loadable(lazy(() => import('../pages/ComingSoon')));